@@ -1,105 +1,38 @@
 {
-  "name": "blk-design-system-angular",
-  "version": "1.2.0",
+  "name": "m1p10mean-mendrika-haendel",
+  "version": "0.0.0",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
     "build": "ng build",
-    "description": "BLK Design System Angular: Free Angular Design System For Bootstrap 4 and Angular",
-    "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "ng e2e",
-    "install:clean": "rm -rf node_modules/ && rm -rf package-lock.json && npm install && npm start"
+    "watch": "ng build --watch --configuration development",
+    "test": "ng test"
   },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/creativetimofficial/blk-design-system-angular.git"
-  },
-  "keywords": [
-    "design system",
-    "angular design system",
-    "bootstrap 4 design system",
-    "angular design system",
-    "bootstrap 4 design",
-    "angular design",
-    "bootstrap 4 system",
-    "angular system",
-    "angular",
-    "bootstrap 4 ui kit",
-    "angular uit kit",
-    "bootstrap 4 kit",
-    "angular kit",
-    "blk",
-    "black",
-    "black angular",
-    "black kit",
-    "blk ui kit",
-    "blk angular ui kit",
-    "creative tim",
-    "creative tim angular",
-    "html kit",
-    "angular kit",
-    "html css template",
-    "angular css template",
-    "web template",
-    "angular template",
-    "bootstrap",
-    "bootstrap 4",
-    "css3 template",
-    "frontend",
-    "angular frontend",
-    "responsive bootstrap template",
-    "responsive angular template",
-    "bootstrap ui kit",
-    "responsive ui kit",
-    "responsive angular ui kit",
-    "blk design system",
-    "blk design system angular"
-  ],
-  "author": "Creative Tim <hello@creative-tim.com> (https://www.creative-tim.com/)",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/creativetimofficial/blk-design-system-angular/issues"
-  },
-  "homepage": "https://demos.creative-tim.com/blk-design-system-angular/",
-  "private": false,
+  "private": true,
   "dependencies": {
-    "@angular/animations": "11.0.5",
-    "@angular/common": "11.0.5",
-    "@angular/compiler": "11.0.5",
-    "@angular/core": "11.0.5",
-    "@angular/forms": "11.0.5",
-    "@angular/platform-browser": "11.0.5",
-    "@angular/platform-browser-dynamic": "11.0.5",
-    "@angular/router": "11.0.5",
-    "chart.js": "2.9.4",
-    "core-js": "3.8.2",
-    "jw-bootstrap-switch-ng2": "2.0.5",
-    "ngx-bootstrap": "6.2.0",
-    "nouislider": "14.6.3",
-    "rxjs": "6.6.3",
-    "tslib": "2.0.3",
-    "zone.js": "0.11.3"
+    "@angular/animations": "^15.1.0",
+    "@angular/common": "^15.1.0",
+    "@angular/compiler": "^15.1.0",
+    "@angular/core": "^15.1.0",
+    "@angular/forms": "^15.1.0",
+    "@angular/platform-browser": "^15.1.0",
+    "@angular/platform-browser-dynamic": "^15.1.0",
+    "@angular/router": "^15.1.0",
+    "rxjs": "~7.8.0",
+    "tslib": "^2.3.0",
+    "zone.js": "~0.12.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.1100.5",
-    "@angular/cli": "11.0.5",
-    "@angular/compiler-cli": "11.0.5",
-    "@angular/language-service": "11.0.5",
-    "@types/node": "14.14.20",
-    "@types/jasmine": "3.6.2",
-    "@types/jasminewd2": "2.0.8",
-    "codelyzer": "6.0.1",
-    "jasmine-core": "3.6.0",
-    "jasmine-spec-reporter": "6.0.0",
-    "karma": "5.2.3",
-    "karma-chrome-launcher": "3.1.0",
-    "karma-coverage-istanbul-reporter": "3.0.3",
-    "karma-jasmine": "4.0.1",
-    "karma-jasmine-html-reporter": "1.5.4",
-    "protractor": "7.0.0",
-    "ts-node": "9.1.1",
-    "tslint": "6.1.3",
-    "typescript": "4.0.3"
+    "@angular-devkit/build-angular": "^15.1.1",
+    "@angular/cli": "~15.1.1",
+    "@angular/compiler-cli": "^15.1.0",
+    "@types/jasmine": "~4.3.0",
+    "jasmine-core": "~4.5.0",
+    "karma": "~6.4.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage": "~2.2.0",
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "~2.0.0",
+    "typescript": "~4.9.4"
   }
 }
