@@ -1,29 +1,7 @@
-/*
-=========================================================
-* BLK Design System Angular - v1.2.0
-=========================================================
+import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
 
-* Product Page: https://www.creative-tim.com/product/blk-design-system-angular
-* Copyright 2021 Creative Tim (https://www.creative-tim.com)
-* Licensed under MIT
+import { AppModule } from './app/app.module';
 
-* Coded by Creative Tim
 
-=========================================================
-
-* The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
- */
-
-import { enableProdMode } from "@angular/core";
-import { platformBrowserDynamic } from "@angular/platform-browser-dynamic";
-
-import { AppModule } from "./app/app.module";
-import { environment } from "./environments/environment";
-
-if (environment.production) {
-  enableProdMode();
-}
-
-platformBrowserDynamic()
-  .bootstrapModule(AppModule)
+platformBrowserDynamic().bootstrapModule(AppModule)
   .catch(err => console.error(err));
